requirements.txt
streamlit>=1.38.0
indic-nlp-library
scikit-learn
SpeechRecognition
